+++ src/modules/admin/widgets/navs/Submenu.php
@@ -72,7 +72,7 @@ class Submenu extends \davidhirtz\yii2\skeleton\widgets\fontawesome\Submenu
 
         if (!$this->title) {
             $this->title = $isEntry ? Html::a($model->getI18nAttribute('name'), ['/admin/entry/update', 'id' => $model->id]) :
-                Html::a($this->getParentModule()->name, $this->getParentModule()->url);
+                Html::a($this->getParentModule()->name, $this->getParentModule()->route);
         }
 
         if ($this->title && $this->showUrl) {